refactor: Parser 패키지 변경
- format에서 http.parser로 이동
