+++ src/main/java/codesquad/http/parser/HttpHeadersParser.java
@@ -1,6 +1,5 @@
 package codesquad.http.parser;
 
-import codesquad.format.Parser;
 import codesquad.http.HttpHeaders;
 
 import java.util.Arrays;
